A pre-cut list of some favourite shaders from shadertoy.com
Shadertoy.com lets people play with and exchange GLSL shaders in the browser.  It's big in the demoscene and some people build some amazing animations with ray marching / ray tracing and a few lines of GLSL shader.  This list uses the embed link for some of the best.  Maybe it would be better to add a random shadertoy link selector.  Caution some of these are highly GPU intensive at high resolutions.